feat: rework Ravenous Claw (2nd revision, blood charge)
- increases bleed chance from 0%/6%/12% to 0%/8%/20%
- decreases attack damage to 1/3/5
- decreases attack speed from 5 to 4

+++ src/main/java/com/github/elenterius/biomancy/item/livingtool/LivingClawsItem.java
@@ -35,7 +35,7 @@ public class LivingClawsItem extends SimpleClawsItem implements LivingTool, IKey
 
 	private final int maxNutrients;
 
-	public LivingClawsItem(Tier tier, int baseAttackDamage, float attackSpeedModifier, float attackRangeModifier, int maxNutrients, Properties properties) {
+	public LivingClawsItem(Tier tier, float baseAttackDamage, float attackSpeedModifier, float attackRangeModifier, int maxNutrients, Properties properties) {
 		super(tier, baseAttackDamage, attackSpeedModifier, attackRangeModifier, properties);
 		this.maxNutrients = maxNutrients;
 	}
