+++ src/main/java/com/github/elenterius/biomancy/item/weapon/SimpleClawsItem.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 public class SimpleClawsItem extends ClawsItem implements ICustomTooltip {
 
-	public SimpleClawsItem(Tier tier, int baseAttackDamage, float attackSpeedModifier, float attackRangeModifier, Properties properties) {
+	public SimpleClawsItem(Tier tier, float baseAttackDamage, float attackSpeedModifier, float attackRangeModifier, Properties properties) {
 		super(tier, baseAttackDamage, attackSpeedModifier, attackRangeModifier, properties);
 	}
